chore: Tweak export-ignore configuration to exclude top-level LICENSE, icon.png, dotfiles
The LICENSE is already included in `addons/asset_placer/` to ship in user
projects. The top-level one could conflict with user projects, or appear
to define their license terms.

Ensure we don't export dotfiles like `.gitignore` or `.gitattributes`
in archives either, and remove stray `.DS_Store` files.

Remove usage of top-level icon.svg as placeholder in the plugin, replace it
by in-plugin icon.png with the same size (128x128).